+++ packages/browser-wallet/CHANGELOG.md
@@ -4,6 +4,7 @@
 
 ### Fixed
 
+-   Validate during Baker/delegation stake reduction that at disposal funds can cover the fee.
 -   Delegation target being changed to passive delegation when the user did not choose to update it.
 
 ## 1.0.3